+++ internal/applets/applet.go
@@ -23,6 +23,7 @@ import (
 	"sort"
 	"strconv"
 
+	"github.com/nao1215/mimixbox/internal/applets/fileutils/chgrp"
 	"github.com/nao1215/mimixbox/internal/applets/fileutils/cp"
 	"github.com/nao1215/mimixbox/internal/applets/fileutils/ln"
 	"github.com/nao1215/mimixbox/internal/applets/fileutils/mkdir"
@@ -86,6 +87,7 @@ func init() {
 		"basename":  {basename.Run, "Print basename (PATH without\"/\") from file path"},
 		"cat":       {cat.Run, "Concatenate files and print on the standard output"},
 		"cowsay":    {cowsay.Run, "Print message with cow's ASCII art"},
+		"chgrp":     {chgrp.Run, "Change the group of each FILE to GROUP"},
 		"chroot":    {chroot.Run, "Run command or interactive shell with special root directory"},
 		"cp":        {cp.Run, "Copy file(s) otr Directory(s)"},
 		"dirname":   {dirname.Run, "Print only directory path"},